Feat: add auth.js

+++ controller/LikedController.js
@@ -1,3 +1,4 @@
+const authorizationJwt = require('../auth');
 const jwt = require('jsonwebtoken');
 const conn = require('../mariadb');
 const {StatusCodes} = require('http-status-codes');
@@ -7,51 +8,62 @@ dotenv.config();
 const addLikes = (req, res) => {
   const book_id = req.params.id;
 
-  let authorization = authorizationJwt(req);
+  let authorization = authorizationJwt(req, res);
 
-  let sql = 'INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)';
-  let values = [authorization.id, book_id];
-
-  conn.query(sql, values,
-    (err, results) => {
-      if (err) {
-        console.log(err);
-        return res.status(StatusCodes.BAD_REQUEST).end();
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하시오."
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      "message" : "잘못된 토큰입니다."
+    });
+  } else {
+    let sql = 'INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)';
+    let values = [authorization.id, book_id];
+  
+    conn.query(sql, values,
+      (err, results) => {
+        if (err) {
+          console.log(err);
+          return res.status(StatusCodes.BAD_REQUEST).end();
+        }
+  
+        return res.status(StatusCodes.OK).json(results);
       }
-
-      return res.status(StatusCodes.OK).json(results);
-    }
-  );
+    );
+  }
 }
 
 const deleteLikes = (req, res) => {
   const book_id = req.params.id;
 
-  let authorization = authorizationJwt(req);
+  let authorization = authorizationJwt(req, res);
 
-  let sql = 'DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?';
-  let values = [authorization.id, book_id];
-
-  conn.query(sql, values,
-    (err, results) => {
-      if (err) {
-        console.log(err);
-        return res.status(StatusCodes.BAD_REQUEST).end();
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하시오."
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      "message" : "잘못된 토큰입니다."
+    });
+  } else {
+    let sql = 'DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?';
+    let values = [authorization.id, book_id];
+  
+    conn.query(sql, values,
+      (err, results) => {
+        if (err) {
+          console.log(err);
+          return res.status(StatusCodes.BAD_REQUEST).end();
+        }
+  
+        return res.status(StatusCodes.OK).json(results);
       }
+    );
+  }
 
-      return res.status(StatusCodes.OK).json(results);
-    }
-  );
-}
-
-function authorizationJwt (req) {
-  let receiveJwt = req.headers["authorization"];
-  console.log(receiveJwt);
-
-  let decodedJwt = jwt.verify(receiveJwt, process.env.PRIVATE_KEY);
-  console.log(decodedJwt);
-
-  return decodedJwt;
 }
 
 module.exports = {
